# CW1/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Phantasmagoria Alumni Engagement Platform API
  description: |
    # How It Works
    1. **Alumni as Influencers**: Successful alumni leverage their professional credentials on the platform.
    2. **Sponsorship Model**: Course providers, licensing bodies, and certification organisations sponsor alumni to promote their offerings. Sponsors offer money to alumni for endorsing their credentials. Sponsors are **only charged if the alumni wins** and their profile is displayed. Alumni pocket their winning bid amount.
    3. **Daily Blind Bidding**: Alumni use **sponsorship funds** to bid for exclusive 24-hour featured slots ("Alumni of the Day"). The bid amount cannot exceed the alumni's available sponsorship balance. Alumni keep the difference between bid amount and sponsorship as profit.
    4. **Student Benefits**: Students discover relatable success stories, career paths, and LinkedIn connections.

    ## Authentication
    - **Alumni Users**: Cookie-based session authentication after email verification.
    - **API Clients**: Bearer API key authentication for programmatic access to public endpoints.

    Alumni flow:
    1. `POST /auth/sessions` to log in.
    2. Server sets `access_token` (HttpOnly) and `csrf_token` cookies.
    3. For authenticated write operations (`POST`, `PUT`, `PATCH`, `DELETE`), send the `csrf_token` cookie and an `X-CSRF-Token` header with the same value.
...

    ## Bidding Rules
...

    ## Rate Limiting
...

  version: 1.0.0
...
servers:
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        Alumni authenticated session cookie.
        Issued by `POST /auth/sessions` as `Set-Cookie: access_token=...`.
        Used for non-public alumni/admin/sponsor endpoints.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: |
        CSRF cookie required for mutating operations.
        Its value must match the `X-CSRF-Token` header.
        Safe routes do not require this cookie and usually issue it automatically if missing.
        `GET /auth/csrf-token` can be used to fetch it explicitly.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF header required for mutating operations.
        Must match the `csrf_token` cookie value.
        `POST /auth/sessions` is exempt from CSRF validation.
    APIKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API client token for accessing public endpoints.
        Include in the `Authorization` header as `Bearer <token>`.
        API client tokens are scoped to viewing public data.
    

  # ──────────────────────────────────────────
  # Reusable Schemas
  # ──────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred.
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ─── Registration ─────────────────────
...
    RequestEmailVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Alumni email address to send a fresh verification link.
          example: jane.doe@eastminster.ac.uk

    # ─── Login ────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jane.doe@eastminster.ac.uk
        password:
          type: string
          format: password
          example: S3cure!Pass
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful.
        data:
          type: object
          properties:
            token:
              type: string
              description: Signed JWT value stored in the `access_token` cookie. Returned for compatibility but cookie auth is authoritative.
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NjVhMWIyYzNkNGU1ZjZhN2I4YzlkMGUiLCJpYXQiOjE3MDk4MzA0MDAsImV4cCI6MTcwOTgzNDAwMH0.abc123
            expiresIn:
              type: integer
              description: Token expiry time in seconds.
              example: 3600
            userId:
              type: string
              example: 665a1b2c3d4e5f6a7b8c9d0e
            email:
              type: string
              example: jane.doe@eastminster.ac.uk
            firstName:
              type: string
              example: Jane
            lastName:
              type: string
              example: Doe
    SessionSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            activeSessionsCount:
              type: integer
              example: 2
            lastLoginAt:
              type: string
              format: date-time
              nullable: true
              example: 2026-03-02T18:00:00.000Z

    # ─── Password Reset ──────────────────
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: jane.doe@eastminster.ac.uk
    ResetPasswordRequest:
      type: object
      required: [token, newPassword, confirmNewPassword]
      properties:
        token:
          type: string
          description: Secure password reset token sent via email. Single-use with expiry.
          example: f1e2d3c4b5a69788796a5b4c3d2e1f00
        newPassword:
          type: string
          format: password
          minLength: 8
          example: N3wS3cure!Pass
        confirmNewPassword:
          type: string
          format: password
          example: N3wS3cure!Pass

    # ─── Profile ──────────────────────────
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
  responses:
    BadRequest:
      description: Bad request – validation error or malformed input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Validation failed.
            errors:
              - field: email
                message: Must be a valid @eastminster.ac.uk email address.
    Unauthorized:
      description: Unauthorized – missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Authentication required. Please log in.
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Too many requests. Please try again later.
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: An unexpected error occurred. Please try again later.

# ════════════════════════════════════════════
# PATHS
# ════════════════════════════════════════════
paths:
...
  /auth/csrf-token:
    get:
      tags: [Authentication]
      summary: Issue a CSRF token for subsequent mutating requests
      security: []
      description: |
        Issues a CSRF token for clients (e.g. Postman, Swagger UI setup) before calling mutating endpoints.
        Most safe routes also set a `csrf_token` cookie automatically when one is missing.
        Returns the token in response body and sets the `csrf_token` cookie.
      operationId: createCsrfToken
      responses:
        '200':
          description: CSRF token issued.
          headers:
            Set-Cookie:
              description: Sets `csrf_token` cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: CSRF token issued.
                  data:
                    type: object
                    properties:
                      csrfToken:
                        type: string
                        example: 6f79e5f0f1cc4a42a13fd58fdf7d98f95ca22a6d4a6428a5795e22692e85d07f
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/verify-email/{token}:
    get:
      tags: [Authentication]
      summary: Verify alumni email address
      security: []
      description: |
        Verifies the alumni's email address using the token sent during registration.
        The token is **cryptographically random**, **single-use**, and has a configurable
        **expiry time** (default: 24 hours). Once verified, the alumni can log in.
      operationId: verifyEmail
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: Email verification token from the verification email.
          example: a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6
      responses:
        '200':
          description: Email verified successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Email verified successfully. You can now log in.
        '400':
          description: Invalid or expired verification token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Invalid or expired verification token. Please request a new one.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/verify-email:
    post:
      tags: [Authentication]
      summary: Request a new email verification token
      security:
        - CsrfCookie: []
          CsrfHeader: []
      description: |
        Creates a new verification token and sends the verification email. The previous token is invalidated.
        This is a mutating endpoint and requires CSRF (`csrf_token` cookie + `X-CSRF-Token` header).
      operationId: createEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEmailVerificationRequest'
      responses:
        '201':
          description: Verification token created and email sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Verification email has been sent. Please check your inbox.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/sessions:
    post:
      tags: [Authentication]
      summary: Create a new authentication session
      security: []
      description: |
        Authenticates the alumni and creates a cookie-based session.
        On success, the server sets:
        - `access_token` (HttpOnly) for authentication
        - `csrf_token` (readable cookie) for CSRF protection

        Login is blocked if the email has not been verified.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Session created. Authentication and CSRF cookies are set.
          headers:
            Set-Cookie:
              description: Sets `access_token` and `csrf_token` cookies.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials or email not verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Invalid email or password.
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Authentication]
      summary: Destroy the current authentication session
      description: Invalidates the current session and clears `access_token` and `csrf_token` cookies.
      operationId: deleteSession
      security:
        - SessionCookieAuth: []
          CsrfCookie: []
          CsrfHeader: []
      responses:
        '204':
          description: Session destroyed successfully. No content returned.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/sessions/summary:
    get:
      tags: [Authentication]
      summary: Get current user's session summary
      description: Returns the current user's active session count and last login time.
      operationId: getSessionSummary
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Session summary returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/password-resets:
    post:
      tags: [Authentication]
      summary: Create a password reset request
      security:
        - CsrfCookie: []
          CsrfHeader: []
      description: |
        Creates a password reset resource and sends an email with a **cryptographically random, single-use** token.
        The token expires after a configurable period (default: 1 hour).
        For security, a success response is returned regardless of whether the email exists.
      operationId: createPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '201':
          description: Password reset resource created (email sent if account exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: If an account with that email exists, a password reset token has been sent.
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Authentication]
      summary: Complete a password reset
      security:
        - CsrfCookie: []
          CsrfHeader: []
      description: |
        Completes a password reset using the secure token from the password reset email.
        The token is validated for authenticity, expiry, and single-use.
        The new password must meet the same strength requirements as registration.
        All existing JWT tokens for the user are invalidated (blacklisted) after a successful reset.
      operationId: completePasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Password has been reset successfully. Please log in with your new password.
        '400':
          description: Invalid or expired reset token, or password does not meet requirements.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ──────────────────────────────────────────
  # PROFILE MANAGEMENT
  # ──────────────────────────────────────────
...